# store: optimized data retrieval from Prometheus