## Changes

Use the new Range/Interval query hints to only select data which is
really needed - instead of sending one big query with min/max time,
divide it into more, smaller queries, each requesting for a smaller
amount of time. Only do this where the step is at least four times
larger than the range to avoid sending a bunch of queries that together
request the same data. Via the benchmarks, I've been able to deduce that
the difference must be at least four times as large to make sense. Here is how the
benchmark numbers look like:

```
name                                                                  time/op
QueryRangeInterval/unoptimized_sidecar,_worst_case,_one_big_block-16  7.37ms ± 5%
QueryRangeInterval/unoptimized_sidecar,_small_blocks_with_2h_step-16  37.5ms ± 5%
QueryRangeInterval/optimized_sidecar,_worst_case,_one_big_block-16    6.62ms ±14%
QueryRangeInterval/optimized_sidecar,_small_blocks_with_2h_step-16    2.84ms ±15%

name                                                                  alloc/op
QueryRangeInterval/unoptimized_sidecar,_worst_case,_one_big_block-16  41.3kB ± 1%
QueryRangeInterval/unoptimized_sidecar,_small_blocks_with_2h_step-16  34.2kB ± 1%
QueryRangeInterval/optimized_sidecar,_worst_case,_one_big_block-16    41.1kB ± 1%
QueryRangeInterval/optimized_sidecar,_small_blocks_with_2h_step-16    33.0kB ± 0%

name                                                                  allocs/op
QueryRangeInterval/unoptimized_sidecar,_worst_case,_one_big_block-16     389 ± 0%
QueryRangeInterval/unoptimized_sidecar,_small_blocks_with_2h_step-16     265 ± 0%
QueryRangeInterval/optimized_sidecar,_worst_case,_one_big_block-16       389 ± 0%
QueryRangeInterval/optimized_sidecar,_small_blocks_with_2h_step-16       264 ± 0%
```

In the future, this optimization could be applied to other components that have a TSDB as well
such as Receive/Rule.

* [x] I added CHANGELOG entry for this change.
* [ ] Change is not relevant to the end user.

Signed-off-by: Giedrius Statkevičius <giedrius.statkevicius@vinted.com>

## Verification

Running manually with quickstart, tests pass.