@GiedriusS 2021-11-15
> Is this a way to make sure the query result remains the same? Can we run the query compliance suite in CI and make sure the result is consistent as the last release?

I ran it after fixing the compliance suite (https://github.com/thanos-io/thanos/pull/4873):

```
compatibility_test.go:74: ================================================================================
        General query tweaks:
        *  Thanos requires adding "external_labels" to distinguish Prometheus servers, leading to extra labels in query results that need to be stripped before comparing results.
        ================================================================================
        Total: 529 / 529 (100.00%) passed, 0 unsupported
         112 / 529 [--------------->___________________________________________________________] 21.17% ? p/s174 / 529 [------------------------>__________________________________________________] 32.89% ? p/s268 / 529 [------------------------------------->_____________________________________] 50.66% ? p/s331 / 529 [--------------------------------------------->___________________________] 62.57% 364 p/s372 / 529 [--------------------------------------------------->_____________________] 70.32% 364 p/s480 / 529 [------------------------------------------------------------------>______] 90.74% 364 p/s520 / 529 [----------------------------------------------------------------------->_] 98.30% 361 p/s524 / 529 [------------------------------------------------------------------------>] 99.05% 361 p/s528 / 529 [------------------------------------------------------------------------>] 99.81% 361 p/s529 / 529 [------------------------------------------------------------------------] 100.00% 326 p/s
=== RUN   TestPromQLCompliance/sidecar
    compatibility_test.go:82: ================================================================================
        General query tweaks:
        *  Thanos requires adding "external_labels" to distinguish Prometheus servers, leading to extra labels in query results that need to be stripped before comparing results.
        ================================================================================
        Total: 529 / 529 (100.00%) passed, 0 unsupported
         125 / 529 [----------------->_________________________________________________________] 23.63% ? p/s186 / 529 [-------------------------->________________________________________________] 35.16% ? p/s289 / 529 [---------------------------------------->__________________________________] 54.63% ? p/s355 / 529 [------------------------------------------------>________________________] 67.11% 383 p/s431 / 529 [----------------------------------------------------------->_____________] 81.47% 383 p/s519 / 529 [----------------------------------------------------------------------->_] 98.11% 383 p/s523 / 529 [------------------------------------------------------------------------>] 98.87% 377 p/s527 / 529 [------------------------------------------------------------------------>] 99.62% 377 p/s529 / 529 [------------------------------------------------------------------------] 100.00% 361 p/s
```