@brian-brazil 2021-11-15
From a quick peek this looks safe from a correctness standpoint, as the discarded data is never used by PromQL. However I would wonder if the queries this helps with are reasonable queries in the first place. If you're not covering all of the range, then the result of e.g. `rate()` could be highly misleading as you'd be ignoring at least 75% of data. I'd suggest that it's not worth spending complexity on optimising bad queries.