@yeya24 2021-12-12
@GiedriusS Would you still like to keep working on this or we just switch to the pushdown approach?